Added missing SurfaceMaterial::copy

+++ src/tests/SpaceSegment_Test.cpp
@@ -21,7 +21,7 @@ TEST(SpaceSegment, iterateOnGrid)
     SpaceSegment segment(Vector3(0.5, 1.5, 0.0), Vector3(2.5, 0.5, 0.0));
     segment.iterateOnGrid(it);
 
-    ASSERT_EQ(4, it.locations.size());
+    ASSERT_EQ(4, (int)it.locations.size());
     EXPECT_VECTOR3_COORDS(it.locations[0], 0.0, 1.0, 0.0);
     EXPECT_VECTOR3_COORDS(it.locations[1], 1.0, 1.0, 0.0);
     EXPECT_VECTOR3_COORDS(it.locations[2], 1.0, 0.0, 0.0);
@@ -34,7 +34,7 @@ TEST(SpaceSegment, iterateOnGrid_Corner)
     SpaceSegment segment(Vector3(0.5, 0.5, 0.5), Vector3(2.5, 2.5, 2.5));
     segment.iterateOnGrid(it);
 
-    ASSERT_EQ(7, it.locations.size());
+    ASSERT_EQ(7, (int)it.locations.size());
     EXPECT_VECTOR3_COORDS(it.locations[0], 0.0, 0.0, 0.0);
     EXPECT_VECTOR3_COORDS(it.locations[3], 1.0, 1.0, 1.0);
     EXPECT_VECTOR3_COORDS(it.locations[6], 2.0, 2.0, 2.0);
@@ -46,7 +46,7 @@ TEST(SpaceSegment, iterateOnGrid_OneCell)
     SpaceSegment segment(Vector3(8.1, 8.2, 8.9), Vector3(8.9, 8.3, 8.6));
     segment.iterateOnGrid(it);
 
-    ASSERT_EQ(1, it.locations.size());
+    ASSERT_EQ(1, (int)it.locations.size());
     EXPECT_VECTOR3_COORDS(it.locations[0], 8.0, 8.0, 8.0);
 }
 
@@ -56,6 +56,6 @@ TEST(SpaceSegment, iterateOnGrid_Negative)
     SpaceSegment segment(Vector3(-8.1, -8.2, -8.9), Vector3(-8.9, -8.3, -8.6));
     segment.iterateOnGrid(it);
 
-    ASSERT_EQ(1, it.locations.size());
+    ASSERT_EQ(1, (int)it.locations.size());
     EXPECT_VECTOR3_COORDS(it.locations[0], -9.0, -9.0, -9.0);
 }
